requirements.txt
streamlit
tensorflow
scikit-learn
numpy
pandas
matplotlib
Pillow
joblib
librosa
soundfile
